+++ src/taurenmd/cli_nosol.py
@@ -25,8 +25,9 @@ Remove all solvent except for NA atoms:
 import argparse
 import functools
 
-import taurenmd.core as tcore
-from taurenmd import _BANNER, log
+from taurenmd import _BANNER
+from taurenmd import core as tcore
+from taurenmd import log
 from taurenmd.libs import libcli, libio, libmdt
 from taurenmd.logger import S, T
 
@@ -49,7 +50,7 @@ ap = libcli.CustomParser(
 
 libcli.add_version_arg(ap)
 libcli.add_topology_arg(ap)
-libcli.add_trajectory_arg(ap)
+libcli.add_trajectories_arg(ap)
 libcli.add_traj_output_arg(ap)
 libcli.add_top_output_arg(ap)
 
@@ -71,7 +72,7 @@ def _ap():
 
 def main(
         topology,
-        trajectory,
+        trajectories,
         selection=None,
         maintain=None,
         top_output='_frame0.pdb',
@@ -81,20 +82,20 @@ def main(
     """Execute main client logic."""
     log.info(T('Removing solvent'))
 
-    trj = libmdt.load_traj(topology, trajectory)
-    
+    trj = libmdt.load_traj(topology, trajectories)
+
     if selection:
         log.info(S('selecting {}', selection))
         atom_sel = trj.top.select(selection)
         trj.atom_slice(atom_sel, inplace=True)
 
     trj.remove_solvent(inplace=True, exclude=maintain)
-    
+
     if top_output:
         fout = libio.parse_top_output(top_output, traj_output)
         trj[0].save(fout.str())
         log.info(S('first frame topology saved: {}', fout))
-   
+
     log.info(T('saving trajectory'))
     log.info(S('destination: {}', traj_output))
     trj.save(traj_output)