+++ src/taurenmd/cli_rmsf.py
@@ -40,8 +40,9 @@ import argparse
 import functools
 from datetime import datetime
 
-import taurenmd.core as tcore
-from taurenmd import _BANNER, Path, log  # noqa: F401
+from taurenmd import _BANNER, Path  # noqa: F401
+from taurenmd import core as tcore
+from taurenmd import log
 from taurenmd.libs import libcalc, libcli, libio, libmda, libplot
 from taurenmd.logger import S, T